requirements.txt
astroid==2.2.5
certifi==2019.6.16
Django==2.2.5
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-jsignature==0.8
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pylint==2.3.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.1
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
